pyproject.toml
[tool.poetry]
name = "simulation-starter-kit"
version = "0.1.0"
description = "A Python SimPy simulation starter kit"
authors = ["Harry Munro <harry.j.munro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
simpy = "^4.0"
pandas = "2.2"
pyyaml = "^6.0"
pytest = "^7.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^3.0"
pytest-bdd = "^4.0"